requirements.txt
a2wsgi==1.10.7
absl-py==2.1.0
aiohappyeyeballs==2.4.3
aiohttp==3.11.4
aiosignal==1.3.1
aiosqlite==0.20.0
annotated-types==0.7.0
anyio==4.6.2.post1
appdirs==1.4.4
asgiref==3.8.1
asttokens==2.4.1
async-timeout==5.0.1
attrs==24.2.0
bentoml==1.3.13
cattrs==23.1.2
certifi==2024.8.30
charset-normalizer==3.4.0
circus==0.18.0
click==8.1.7
click-option-group==0.5.6
cloudpickle==3.1.0
colorama==0.4.6
coloredlogs==15.0.1
comm==0.2.2
contourpy==1.3.0
cycler==0.12.1
debugpy==1.8.7
decorator==5.1.1
deepmerge==2.0
Deprecated==1.2.15
exceptiongroup==1.2.2
executing==2.1.0
filelock==3.16.1
flatbuffers==24.3.25
fonttools==4.54.1
frozenlist==1.5.0
fs==2.4.16
fsspec==2024.10.0
grpcio==1.67.1
h11==0.14.0
httpcore==1.0.7
httpx==0.27.2
httpx-ws==0.6.2
humanfriendly==10.0
idna==3.10
importlib-metadata==6.11.0
importlib_resources==6.4.5
inflection==0.5.1
ipykernel==6.29.5
ipython==8.18.1
jedi==0.19.1
Jinja2==3.1.4
joblib==1.4.2
jupyter_client==8.6.3
jupyter_core==5.7.2
kiwisolver==1.4.7
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.9.2
matplotlib-inline==0.1.7
mdurl==0.1.2
mpmath==1.3.0
multidict==6.1.0
nest-asyncio==1.6.0
networkx==3.2.1
numpy==1.26.4
nvidia-ml-py==12.560.30
onnx==1.17.0
onnx-simplifier==0.4.36
onnxruntime==1.20.0
onnxslim==0.1.39
opencv-python==4.10.0.84
opentelemetry-api==1.20.0
opentelemetry-instrumentation==0.41b0
opentelemetry-instrumentation-aiohttp-client==0.41b0
opentelemetry-instrumentation-asgi==0.41b0
opentelemetry-sdk==1.20.0
opentelemetry-semantic-conventions==0.41b0
opentelemetry-util-http==0.41b0
packaging==24.1
pandas==2.2.3
parso==0.8.4
pathspec==0.12.1
pillow==11.0.0
pip-requirements-parser==32.0.1
platformdirs==4.3.6
prometheus_client==0.21.0
prompt-toolkit==3.0.36
propcache==0.2.0
protobuf==5.28.3
psutil==6.1.0
pure_eval==0.2.3
py-cpuinfo==9.0.0
pydantic==2.9.2
pydantic_core==2.23.4
Pygments==2.18.0
pyparsing==3.2.0
pyreadline3==3.5.4
python-dateutil==2.9.0.post0
python-json-logger==2.0.7
python-multipart==0.0.17
pytz==2024.2
pywin32==308
PyYAML==6.0.2
pyzmq==26.2.0
questionary==2.0.1
requests==2.32.3
rich==13.9.4
schema==0.7.7
scikit-learn==1.5.2
scipy==1.13.1
seaborn==0.13.2
simple-di==0.1.5
six==1.16.0
sniffio==1.3.1
stack-data==0.6.3
starlette==0.41.3
sympy==1.13.1
tensorboard==2.18.0
tensorboard-data-server==0.7.2
threadpoolctl==3.5.0
tomli==2.1.0
tomli_w==1.1.0
torch==2.0.0+cu117
torchaudio==2.0.1+cu117
torchvision==0.15.1+cu117
tornado==6.4.1
tqdm==4.66.6
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2024.2
ultralytics==8.3.27
ultralytics-thop==2.0.10
urllib3==2.2.3
uv==0.5.2
uvicorn==0.32.0
watchfiles==0.24.0
wcwidth==0.2.13
Werkzeug==3.1.1
wrapt==1.16.0
wsproto==1.2.0
yarl==1.17.2
zipp==3.20.2
